+++ src/main/java/com/jcloisterzone/game/state/GameState.java
@@ -11,7 +11,9 @@ import com.jcloisterzone.feature.Feature;
 import com.jcloisterzone.figure.Meeple;
 import com.jcloisterzone.game.Capability;
 import com.jcloisterzone.game.Rule;
+import com.jcloisterzone.game.capability.AbbeyCapability;
 import com.jcloisterzone.game.capability.CountCapability;
+import com.jcloisterzone.game.phase.AbbeyEndGamePhase;
 import com.jcloisterzone.game.phase.CocFinalScoringPhase;
 import com.jcloisterzone.game.phase.GameOverPhase;
 import com.jcloisterzone.game.phase.Phase;
@@ -370,10 +372,12 @@ public class GameState implements ActionsMixin, BoardMixin,
     }
 
     public Class<? extends Phase> getEndPhase () {
+        if (hasCapability(AbbeyCapability.class)) {
+            return AbbeyEndGamePhase.class;
+        }
         if (hasCapability(CountCapability.class)) {
             return CocFinalScoringPhase.class;
-        } else {
-            return GameOverPhase.class;
         }
+        return GameOverPhase.class;
     }
 }